+++ pkg/cfp/service.go
@@ -9,7 +9,7 @@ import (
 
 type Service interface {
 	CreateOrUpdate(ctx context.Context, cfp Cfp) (result Cfp, err error)
-	DeleteByEventID(ctx context.Context, eventID uint) (err error)
+	DeleteByEventID(ctx context.Context, eventID int64) (err error)
 }
 
 type CfpService struct {
@@ -45,7 +45,7 @@ func (s *CfpService) CreateOrUpdate(ctx context.Context, cfp Cfp) (result Cfp, e
 	return
 }
 
-func (s *CfpService) DeleteByEventID(ctx context.Context, eventID uint) (err error) {
+func (s *CfpService) DeleteByEventID(ctx context.Context, eventID int64) (err error) {
 	if err = s.db.WithContext(ctx).Where("event_id = ?", eventID).Delete(&Cfp{}).Error; err != nil {
 		slog.ErrorContext(ctx, "Fail to delete CFP", "event_id", eventID, "error", err.Error())
 	}
